+++ EasyDriver.Tests/Integration/Ui/LabeledComponentTests.cs
@@ -6,5 +6,5 @@ namespace EasyDriver.Tests.Integration.Ui;
 public class LabeledComponentTests : IntegrationBase {
     public LabeledComponentTests(ITestOutputHelper output, IntegrationFixture fix) : base(output, fix) { }
 
-    [Fact] void LabeledTest() {}
+    [Fact] void LabeledTest() { }
 }