+++ EasyDriver.Tests/Unit/ConfigurationTest.cs
@@ -1,5 +1,4 @@
-﻿
-namespace EasyDriver.Tests.Unit;
+﻿namespace EasyDriver.Tests.Unit;
 
 public class ConfigurationTest {
     [Fact] void ReadConfigTest() {