+++ EasyDriver/EasyDriver/Se/Infra/Browser/SimpleBrowserRunner.cs
@@ -5,8 +5,10 @@ namespace Comfast.EasyDriver.Se.Infra.Browser;
 
 internal class SimpleBrowserRunner : IBrowserRunner {
     private readonly Func<IWebDriver> _runFunc;
+
     public SimpleBrowserRunner(Func<IWebDriver> runFunc) {
         _runFunc = runFunc;
     }
+
     public IWebDriver RunNewBrowser() => _runFunc();
 }